Correct syntax error and add renewable generator tasks
Fixed a syntax error in the assignment of the B matrix. Removed the comment about the dcline prop from an external file and added a set of tasks related to adding renewable generators to the gen matrix, along with corresponding properties. The paths in the multiperiod_opf_new_with_renewable notebook were also updated. A figure error was encountered due to the missing 'supxlabel' attribute.