Adjust handling of input types in NYGrid module
This commit allows for the usage of numpy arrays and pandas dataframes for DC line, ESR and VRE properties in the NYGrid module. Changes include streamlined references to generator and property arrays, improving overall efficiency. In addition, the capabilities of the module have been expanded by establishing processes for VRE generators within the grid model, and adding error handling for incompatible types of input properties.